MAJOR OVERHAUL: I put all the needed files into different folders
Added some files

[regEx.c]
The derivation method is now memory safe (frees the no longer needed memory)

Instead of using the derivation method alone, I added a function called "patternToDFA", which creates the minimum DFA for each pattern and returns it, thus enormously increasing matching speed and decreasing memory usage even further. I plan on deprecating the old "match" method in the future, if this one proves to be as fast (or even faster). A clear advantage of the DFA is the re-usability of it, because instead of deriving a pattern, it stays the same. However, the derivation method is still needed for this.
If the patternToDFA should replace the previous method, I also need to implement a function which recursively deletes the entire tree